extract code and split firework formatters according to type
